refactor: move PaginatedPlatformAssistantsResponse into AuthModels.swift
Now that the pre-commit hook allowlists the webhookSecret CodingKey
false positive (#24145), the paginated list response can live in its
conceptual home next to PlatformAssistant.
